requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Caching>=2.1.0
Pillow>=11.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.1
Werkzeug==3.0.3
requests==2.31.0
email-validator>=2.0.0
google-generativeai>=0.3.0
flask-limiter>=3.5.0
flask-talisman>=1.1.0
flask-migrate>=4.0.0
pytz>=2023.3
redis>=5.0.0
celery>=5.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.2.0
black>=23.0.0
flake8>=6.0.0
pylint>=3.0.0
mypy>=1.7.0
sentry-sdk[flask]>=1.40.0
